Complete numpy array type annotations and fix return type casting
- Add proper np.ndarray[Any, np.dtype[Any]] type parameters to all remaining methods in functional_diversity.py
- Fix np.mean return type casting in quality_assessment.py _calculate_patch_diversity method
- Update _perform_trait_pca and _calculate_trait_summary_stats method signatures
- Ensure all numpy function returns are explicitly cast to expected types

Progress: 357 type errors remaining (down from 363)

+++ src/preprocessing/quality_assessment.py
@@ -222,7 +222,7 @@ class QualityAssessor(LoggerMixin):
         if not correlations:
             return 0.0
         
-        diversity_score = 1.0 - np.mean(correlations)
+        diversity_score = 1.0 - float(np.mean(correlations))
         return max(0.0, diversity_score)
     
     def comprehensive_assessment(self, image: np.ndarray[Any, np.dtype[Any]]) -> Dict[str, Any]:
